+++ CHANGES.md
@@ -25,6 +25,15 @@ Fixes:
 
 ## Current
 
+**2.6.75 - 2025-06-09**
+
+Changes:
+- Update kan and kanctapp scrapers
+- update `nc` scraper to OpinionSiteLinear and new website #1373
+
+
+## Past
+
 **2.6.74 - 2025-06-04**
 
 - Add `test_hashes` optional argument to `sample_caller`. Helpful to detect
@@ -46,8 +55,6 @@ Fixes:
 
 Features:
 - Fix for CA4 - minor edge case bug
-
-## Past
 
 **2.6.69 - 2025-05-21**